Update source to 37.0.1

+++ testing/mochitest/runtests.py
@@ -108,10 +108,15 @@ class MessageLogger(object):
     VALID_ACTIONS = set(['suite_start', 'suite_end', 'test_start', 'test_end',
                          'test_status', 'log',
                          'buffering_on', 'buffering_off'])
+    TEST_PATH_PREFIXES = ['/tests/',
+                          'chrome://mochitests/content/browser/',
+                          'chrome://mochitests/content/chrome/']
+
 
     def __init__(self, logger, buffering=True):
         self.logger = logger
         self.buffering = buffering
+        self.restore_buffering = False
         self.tests_started = False
 
         # Message buffering
@@ -123,6 +128,16 @@ class MessageLogger(object):
     def valid_message(self, obj):
         """True if the given object is a valid structured message (only does a superficial validation)"""
         return isinstance(obj, dict) and 'action' in obj and obj['action'] in MessageLogger.VALID_ACTIONS
+
+    def _fix_test_name(self, message):
+      """Normalize a logged test path to match the relative path from the sourcedir.
+      """
+      if 'test' in message:
+        test = message['test']
+        for prefix in MessageLogger.TEST_PATH_PREFIXES:
+          if test.startswith(prefix):
+            message['test'] = test[len(prefix):]
+            break
 
     def parse_line(self, line):
         """Takes a given line of input (structured or not) and returns a list of structured messages"""
@@ -138,6 +153,7 @@ class MessageLogger(object):
                     message = dict(action='log', level='info', message=fragment, unstructured=True)
             except ValueError:
                 message = dict(action='log', level='info', message=fragment, unstructured=True)
+            self._fix_test_name(message)
             messages.append(message)
 
         return messages
@@ -160,42 +176,43 @@ class MessageLogger(object):
             unstructured = True
             message.pop('unstructured')
 
-        # Saving errors/failures to be shown at the end of the test run
-        is_error = 'expected' in message or (message['action'] == 'log' and message['message'].startswith('TEST-UNEXPECTED'))
-        if is_error:
+        # Error detection also supports "raw" errors (in log messages) because some tests
+        # manually dump 'TEST-UNEXPECTED-FAIL'.
+        if ('expected' in message or
+            (message['action'] == 'log' and message['message'].startswith('TEST-UNEXPECTED'))):
+            # Saving errors/failures to be shown at the end of the test run
             self.errors.append(message)
-
-        # If we don't do any buffering, or the tests haven't started, or the message was unstructured, it is directly logged
-        if not self.buffering or unstructured or not self.tests_started:
-            self.logger.log_raw(message)
-            return
-
-        # If a test ended, we clean the buffer
-        if message['action'] == 'test_end':
-            self.buffered_messages = []
-
-        # Buffering logic; Also supports "raw" errors (in log messages) because some tests manually dump 'TEST-UNEXPECTED-FAIL'
-        if not is_error and message['action'] not in self.BUFFERED_ACTIONS:
-            self.logger.log_raw(message)
-            return
-
-        # test_status messages buffering
-        if is_error:
+            self.restore_buffering = self.restore_buffering or self.buffering
+            self.buffering = False
             if self.buffered_messages:
                 snipped = len(self.buffered_messages) - self.BUFFERING_THRESHOLD
                 if snipped > 0:
-                  self.logger.info("<snipped {0} output lines - "
-                                   "if you need more context, please use "
-                                   "SimpleTest.requestCompleteLog() in your test>"
-                                   .format(snipped))
+                    self.logger.info("<snipped {0} output lines - "
+                                     "if you need more context, please use "
+                                     "SimpleTest.requestCompleteLog() in your test>"
+                                     .format(snipped))
                 # Dumping previously buffered messages
                 self.dump_buffered(limit=True)
 
             # Logging the error message
             self.logger.log_raw(message)
+        # If we don't do any buffering, or the tests haven't started, or the message was
+        # unstructured, it is directly logged.
+        elif any([not self.buffering,
+                  unstructured,
+                  not self.tests_started,
+                  message['action'] not in self.BUFFERED_ACTIONS]):
+            self.logger.log_raw(message)
         else:
             # Buffering the message
             self.buffered_messages.append(message)
+
+        # If a test ended, we clean the buffer
+        if message['action'] == 'test_end':
+            self.buffered_messages = []
+            if self.restore_buffering:
+                self.restore_buffering = False
+                self.buffering = True
 
     def write(self, line):
         messages = self.parse_line(line)
@@ -347,7 +364,7 @@ class MochitestServer(object):
       time.sleep(1)
       i += 1
     else:
-      self._log.error("TEST-UNEXPECTED-FAIL | runtests.py | Timed out while waiting for server startup.")
+      self._log.info("TEST-UNEXPECTED-FAIL | runtests.py | Timed out while waiting for server startup.")
       self.stop()
       sys.exit(1)
 
@@ -503,8 +520,7 @@ class MochitestUtilsMixin(object):
     if options.logFile:
       options.logFile = self.getLogFilePath(options.logFile)
 
-    # Note that all tests under options.subsuite need to be browser chrome tests.
-    if options.browserChrome or options.chrome or options.subsuite or \
+    if options.browserChrome or options.chrome or \
        options.a11y or options.webapprtChrome or options.jetpackPackage or \
        options.jetpackAddon:
       self.makeTestConfig(options)
@@ -671,9 +687,9 @@ class MochitestUtilsMixin(object):
       paths.append(test)
 
     # Bug 883865 - add this functionality into manifestparser
-    with open(os.path.join(SCRIPT_DIR, 'tests.json'), 'w') as manifestFile:
+    with open(os.path.join(SCRIPT_DIR, options.testRunManifestFile), 'w') as manifestFile:
       manifestFile.write(json.dumps({'tests': paths}))
-    options.manifestFile = 'tests.json'
+    options.manifestFile = options.testRunManifestFile
 
     return self.buildTestURL(options)
 
@@ -692,7 +708,7 @@ class MochitestUtilsMixin(object):
       with open(options.pidFile + ".xpcshell.pid", 'w') as f:
         f.write("%s" % self.server._process.pid)
 
-  def startServers(self, options, debuggerInfo):
+  def startServers(self, options, debuggerInfo, ignoreSSLTunnelExts = False):
     # start servers and set ports
     # TODO: pass these values, don't set on `self`
     self.webServer = options.webServer
@@ -710,7 +726,7 @@ class MochitestUtilsMixin(object):
     self.startWebSocketServer(options, debuggerInfo)
 
     # start SSL pipe
-    self.sslTunnel = SSLTunnel(options, logger=self.log)
+    self.sslTunnel = SSLTunnel(options, logger=self.log, ignoreSSLTunnelExts = ignoreSSLTunnelExts)
     self.sslTunnel.buildConfig(self.locations)
     self.sslTunnel.start()
 
@@ -771,6 +787,18 @@ class MochitestUtilsMixin(object):
       dir = "file:///" + dir.replace("\\", "/")
     return dir
 
+  def writeChromeManifest(self, options):
+    manifest = os.path.join(options.profilePath, "tests.manifest")
+    with open(manifest, "w") as manifestFile:
+      # Register chrome directory.
+      chrometestDir = self.getChromeTestDir(options)
+      manifestFile.write("content mochitests %s contentaccessible=yes\n" % chrometestDir)
+
+      if options.testingModulesDir is not None:
+        manifestFile.write("resource testing-common file:///%s\n" %
+          options.testingModulesDir)
+    return manifest
+
   def addChromeToProfile(self, options):
     "Adds MochiKit chrome tests to the profile."
 
@@ -792,19 +820,11 @@ toolbar#nav-bar {
     with open(os.path.join(options.profilePath, "userChrome.css"), "a") as chromeFile:
       chromeFile.write(chrome)
 
-    manifest = os.path.join(options.profilePath, "tests.manifest")
-    with open(manifest, "w") as manifestFile:
-      # Register chrome directory.
-      chrometestDir = self.getChromeTestDir(options)
-      manifestFile.write("content mochitests %s contentaccessible=yes\n" % chrometestDir)
-
-      if options.testingModulesDir is not None:
-        manifestFile.write("resource testing-common file:///%s\n" %
-          options.testingModulesDir)
+    manifest = self.writeChromeManifest(options)
 
     # Call installChromeJar().
     if not os.path.isdir(os.path.join(SCRIPT_DIR, self.jarDir)):
-      self.log.error("TEST-UNEXPECTED-FAIL | invalid setup: missing mochikit extension")
+      self.log.info("TEST-UNEXPECTED-FAIL | invalid setup: missing mochikit extension")
       return None
 
     # Support Firefox (browser), B2G (shell), SeaMonkey (navigator), and Webapp
@@ -857,7 +877,7 @@ overlay chrome://browser/content/browser.xul chrome://mochikit/content/jetpack-a
     return extensions
 
 class SSLTunnel:
-  def __init__(self, options, logger):
+  def __init__(self, options, logger, ignoreSSLTunnelExts = False):
     self.log = logger
     self.process = None
     self.utilityPath = options.utilityPath
@@ -867,6 +887,7 @@ class SSLTunnel:
     self.httpPort = options.httpPort
     self.webServer = options.webServer
     self.webSocketPort = options.webSocketPort
+    self.useSSLTunnelExts = not ignoreSSLTunnelExts
 
     self.customCertRE = re.compile("^cert=(?P<nickname>[0-9a-zA-Z_ ]+)")
     self.clientAuthRE = re.compile("^clientauth=(?P<clientauth>[a-z]+)")
@@ -892,7 +913,7 @@ class SSLTunnel:
         config.write("redirhost:%s:%s:%s:%s\n" %
                      (loc.host, loc.port, self.sslPort, redirhost))
 
-      if option in ('ssl3', 'rc4', 'failHandshake'):
+      if self.useSSLTunnelExts and option in ('ssl3', 'rc4', 'failHandshake'):
         config.write("%s:%s:%s:%s\n" % (option, loc.host, loc.port, self.sslPort))
 
   def buildConfig(self, locations):
@@ -1159,7 +1180,8 @@ class Mochitest(MochitestUtilsMixin):
     if options.browserChrome and options.timeout:
       options.extraPrefs.append("testing.browserTestHarness.timeout=%d" % options.timeout)
     options.extraPrefs.append("browser.tabs.remote.autostart=%s" % ('true' if options.e10s else 'false'))
-    options.extraPrefs.append("browser.tabs.remote.sandbox=%s" % options.contentSandbox)
+    if options.strictContentSandbox:
+        options.extraPrefs.append("security.sandbox.windows.content.moreStrict=true")
 
     # get extensions to install
     extensions = self.getExtensionsToInstall(options)
@@ -1222,7 +1244,7 @@ class Mochitest(MochitestUtilsMixin):
     # TODO: this should really be upstreamed somewhere, maybe mozprofile
     certificateStatus = self.fillCertificateDB(options)
     if certificateStatus:
-      self.log.error("TEST-UNEXPECTED-FAIL | runtests.py | Certificate integration failed")
+      self.log.info("TEST-UNEXPECTED-FAIL | runtests.py | Certificate integration failed")
       return None
 
     return manifest
@@ -1313,9 +1335,10 @@ class Mochitest(MochitestUtilsMixin):
 
   def cleanup(self, options):
     """ remove temporary files and profile """
-    if self.manifest is not None:
+    if hasattr(self, 'manifest') and self.manifest is not None:
       os.remove(self.manifest)
-    del self.profile
+    if hasattr(self, 'profile'):
+        del self.profile
     if options.pidFile != "":
       try:
         os.remove(options.pidFile)
@@ -1386,7 +1409,7 @@ class Mochitest(MochitestUtilsMixin):
       self.log.info("zombiecheck | Checking for orphan process with PID: %d" % processPID)
       if isPidAlive(processPID):
         foundZombie = True
-        self.log.error("TEST-UNEXPECTED-FAIL | zombiecheck | child process %d still alive after shutdown" % processPID)
+        self.log.info("TEST-UNEXPECTED-FAIL | zombiecheck | child process %d still alive after shutdown" % processPID)
         self.killAndGetStack(processPID, utilityPath, debuggerInfo, dump_screen=not debuggerInfo)
 
     return foundZombie
@@ -1559,7 +1582,7 @@ class Mochitest(MochitestUtilsMixin):
       # record post-test information
       if status:
         self.message_logger.dump_buffered()
-        self.log.error("TEST-UNEXPECTED-FAIL | %s | application terminated with exit code %s" % (self.lastTestSeen, status))
+        self.log.info("TEST-UNEXPECTED-FAIL | %s | application terminated with exit code %s" % (self.lastTestSeen, status))
       else:
         self.lastTestSeen = 'Main app process exited normally'
 
@@ -1600,7 +1623,6 @@ class Mochitest(MochitestUtilsMixin):
     """
     self.setTestRoot(options)
     manifest = self.getTestManifest(options)
-
     if manifest:
       # Python 2.6 doesn't allow unicode keys to be used for keyword
       # arguments. This gross hack works around the problem until we
@@ -1621,13 +1643,20 @@ class Mochitest(MochitestUtilsMixin):
           # In the case where we have a single file, we don't want to filter based on options such as subsuite.
           tests = manifest.active_tests(disabled=disabled, options=None, **info)
           for test in tests:
-              if 'disabled' in test:
-                  del test['disabled']
+            if 'disabled' in test:
+              del test['disabled']
+
       else:
-          tests = manifest.active_tests(disabled=disabled, options=options, **info)
+        tests = manifest.active_tests(disabled=disabled, options=options, **info)
+        if len(tests) == 0:
+          tests = manifest.active_tests(disabled=True, options=options, **info)
+
     paths = []
 
     for test in tests:
+      if len(tests) == 1 and 'disabled' in test:
+        del test['disabled']
+
       pathAbs = os.path.abspath(test['path'])
       assert pathAbs.startswith(self.testRootAbs)
       tp = pathAbs[len(self.testRootAbs):].replace('\\', '/').strip('/')
@@ -1694,7 +1723,7 @@ class Mochitest(MochitestUtilsMixin):
         testsToRun = bisect.pre_test(options, testsToRun, status)
         # To inform that we are in the process of bisection, and to look for bleedthrough
         if options.bisectChunk != "default" and not bisection_log:
-            log.info("TEST-UNEXPECTED-FAIL | Bisection | Please ignore repeats and look for 'Bleedthrough' (if any) at the end of the failure list")
+            self.log.info("TEST-UNEXPECTED-FAIL | Bisection | Please ignore repeats and look for 'Bleedthrough' (if any) at the end of the failure list")
             bisection_log = 1
 
       result = self.doTests(options, onLaunch, testsToRun)
@@ -1737,6 +1766,7 @@ class Mochitest(MochitestUtilsMixin):
     options.totalChunks = None
     options.thisChunk = None
     options.chunkByDir = 0
+    result = 1 # default value, if no tests are run.
     inputTestPath = self.getTestPath(options)
     for dir in dirs:
       if inputTestPath and not inputTestPath.startswith(dir):
@@ -1783,9 +1813,11 @@ class Mochitest(MochitestUtilsMixin):
     # TODO: use mozrunner.local.debugger_arguments:
     # https://github.com/mozilla/mozbase/blob/master/mozrunner/mozrunner/local.py#L42
 
-    debuggerInfo = mozdebug.get_debugger_info(options.debugger,
-                                              options.debuggerArgs,
-                                              options.debuggerInteractive)
+    debuggerInfo = None
+    if options.debugger:
+        debuggerInfo = mozdebug.get_debugger_info(options.debugger,
+                                                  options.debuggerArgs,
+                                                  options.debuggerInteractive)
 
     if options.useTestMediaDevices:
       devices = findTestMediaDevices(self.log)
@@ -1820,8 +1852,8 @@ class Mochitest(MochitestUtilsMixin):
       testURL = self.buildTestPath(options, testsToFilter)
 
       # read the number of tests here, if we are not going to run any, terminate early
-      if os.path.exists(os.path.join(SCRIPT_DIR, 'tests.json')):
-        with open(os.path.join(SCRIPT_DIR, 'tests.json')) as fHandle:
+      if os.path.exists(os.path.join(SCRIPT_DIR, options.testRunManifestFile)):
+        with open(os.path.join(SCRIPT_DIR, options.testRunManifestFile)) as fHandle:
           tests = json.load(fHandle)
         count = 0
         for test in tests['tests']:
@@ -1919,7 +1951,7 @@ class Mochitest(MochitestUtilsMixin):
 
     self.message_logger.dump_buffered()
     self.message_logger.buffering = False
-    self.log.error(error_message)
+    self.log.info(error_message)
 
     browserProcessId = browserProcessId or proc.pid
     self.killAndGetStack(browserProcessId, utilityPath, debuggerInfo, dump_screen=not debuggerInfo)
@@ -2117,7 +2149,7 @@ class Mochitest(MochitestUtilsMixin):
 
   def getTestManifest(self, options):
     if isinstance(options.manifestFile, TestManifest):
-        manifest = options.manifestFile
+      manifest = options.manifestFile
     elif options.manifestFile and os.path.isfile(options.manifestFile):
       manifestFileAbs = os.path.abspath(options.manifestFile)
       assert manifestFileAbs.startswith(SCRIPT_DIR)
@@ -2132,6 +2164,8 @@ class Mochitest(MochitestUtilsMixin):
 
       if os.path.exists(masterPath):
         manifest = TestManifest([masterPath], strict=False)
+      else:
+        self._log.warning('TestManifest masterPath %s does not exist' % masterPath)
 
     return manifest
